refactor, __init__ modules and importing
