+++ .gitignore
@@ -1,4 +1,4 @@
-Datasets/COCO
-__*
+datasets/COCO
+__pycache__
 PROB
 mvits_for_class_agnostic_od
